use settings store instead of context
